requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.31.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.17
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pylibmagic==0.2.2
python-decouple==3.7
python-jose==3.3.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
